v5.1.3: Update build-latest-p2pdma-kernel to v5.1.5
Update the build-latest-p2pdma-kernel to use v5.1.5 from the linux
stable repo. Rather than cherry pick patches on top to implment p2pdma
user-space access we apply patches to the stable kernel to implement
this. Note we also fix a couple of bugs associated with patch
application in the underlying script (build-kernel-debrpm). We also
update the config file for this new kernel.

Signed-off-by: Stephen Bates <sbates@raithlin.com>