patches: Add a patches sub-folder
We cleanup the code by adding a patches sub-folder and moving our
p2pdma patch into this folder. We also update the build-kernel-debrpm
to understand patch folders as well as files.

Note that the patches now **only** support v5.6.x series due to
changes in /drivers/pci/probe.c. We will address this in a later
patch.